Add files via upload
The game was created to play Tetris and the 2048 game at the same time. Because of that, you can see the tile numbers on the Tetris blocks but if you run the "Tetris_2048.py" file, you can see that the Tetris part completed but the 2048 part is not working correctly. The project has all required tetrominoes, rotate method, line elimination, and scoring. You can change the images for different login and exit screens. 
tetromino.py --> All blocks required for Tetris.
tile.py --> Part of the code about tile numbers and colors.
game_grid.py --> Settings for game grid
stddraw.py -->The stddraw module defines functions that allow the user to create a drawing.